requirements.txt
python-telegram-bot==13.15
aiohttp==3.9.1
numpy==1.26.2
pandas==2.1.4
python-dotenv==1.0.0
scipy>=1.9.0
pyyaml>=6.0
httpx>=0.28.0
tenacity>=8.2.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0
flask>=2.3.0
notion-client>=2.2.1

